fix(security): float→Decimal for consensus/agent/onboarding/genesis ECU interfaces
consensus/governance.py: ecu_base_costs Dict[str, Decimal]; get_task_fee_ecu() → Decimal. consensus/engine.py: node_stakes Dict[str, Decimal]. agent.py: stake/reward method signatures → Decimal. onboarding.py: credit/repayment → Decimal. exceptions.py: stake/required → Decimal. genesis/work_task.py: ecu_estimate → Optional[Decimal]. Directive 3 compliance per ILC Coding Security Standards (CLAUDE.md). Identified Phase 1117 A11 audit.

float_kill_01_interface_decimal_cleanup

+++ ilc_core/economics/outcome.py
@@ -9,6 +9,7 @@ For how this is used in the economics sandbox and how it might map to future
 genesis primitives, see docs/protocol_econ_surfaces_mvp.md.
 """
 from dataclasses import dataclass
+from decimal import Decimal
 
 @dataclass
 class TaskOutcome:
@@ -21,8 +22,8 @@ class TaskOutcome:
     """
     task_type: str           # e.g. "claim.submit"
     domain: str | None       # e.g. "easy", "medium", "hard", or None
-    stake_spent: float
-    reward_paid: float
+    stake_spent: Decimal
+    reward_paid: Decimal
     success: bool | None     # None if not applicable
